Revamped the home page feature section and refactored it into a commmon component
